website/requirements.txt
Django==3.0.8
djangorestframework